Address Copilot review on #128
- Deduplicate template-expansion logic: PreviewHost now calls
  expandTreatmentFile() instead of inlining fillTemplates() — single
  code path avoids future drift. expandTreatmentFile moved to its own
  module so PreviewHost's import doesn't pull in the js-yaml parser.
- Guard the refresh handler against the preview panel being disposed
  mid-refresh: capture the panel and treatment directory in locals
  before awaits, and re-check previewPanel after each await so the
  non-null-assertion on the outer ref can't throw.
- Remove unnecessary eslint-disable on the contentFns useMemo: the
  deps array already includes both referenced values.
